* Move the main window's code into a new MainWindow class.
