+++ src/windows/MessageList.cpp
@@ -1383,7 +1383,7 @@ void MessageList::OpenAttachment(AttachmentItem* pItem)
 		buffer[_countof(buffer) - 1] = 0;
 		free((void*)urlTStr);
 		int result = MessageBox(
-			g_Hwnd,
+			GetParent(m_hwnd),
 			buffer,
 			GetTextManager()->GetTString(IDS_DANGEROUS_DOWNLOAD_TITLE),
 			MB_ICONWARNING | MB_YESNO
@@ -1936,12 +1936,12 @@ void MessageList::ConfirmOpenLink(const std::string& link)
 		if (LOBYTE(GetVersion()) >= 4)
 		{
 			// TODO: This actually works on NT 3.51 at first.  But the second time this causes an abort
-			if (MessageBoxHooked(g_Hwnd, buffer, TmGetTString(IDS_HOLD_UP_CONFIRM), MB_ICONWARNING | MB_OKCANCEL, IDOK, TmGetTString(IDS_EXCITED_YES)) != IDOK)
+			if (MessageBoxHooked(GetMainHWND(), buffer, TmGetTString(IDS_HOLD_UP_CONFIRM), MB_ICONWARNING | MB_OKCANCEL, IDOK, TmGetTString(IDS_EXCITED_YES)) != IDOK)
 				return;
 		}
 		else
 		{
-			if (MessageBox(g_Hwnd, buffer, TmGetTString(IDS_HOLD_UP_CONFIRM), MB_ICONWARNING | MB_YESNO) != IDYES)
+			if (MessageBox(GetMainHWND(), buffer, TmGetTString(IDS_HOLD_UP_CONFIRM), MB_ICONWARNING | MB_YESNO) != IDYES)
 				return;
 		}
 	}
@@ -3169,11 +3169,11 @@ void MessageList::Paint(HDC hdc, RECT& paintRect)
 
 			if (!m_bManagedByOwner) {
 				if (index >= 100 || hasUnloadedMessagesBelow) {
-					PostMessage(g_Hwnd, WM_SETBROWSINGPAST, 1, 0);
+					PostMessage(GetParent(m_hwnd), WM_SETBROWSINGPAST, 1, 0);
 					sent = true;
 				}
 				else if (!sent && index <= 70) {
-					PostMessage(g_Hwnd, WM_SETBROWSINGPAST, 0, 0);
+					PostMessage(GetParent(m_hwnd), WM_SETBROWSINGPAST, 0, 0);
 				}
 			}
 		}
@@ -3269,7 +3269,7 @@ void MessageList::HandleRightClickMenuCommand(int command)
 		}
 		case ID_DUMMYPOPUP_EDITMESSAGE:
 		{
-			SendMessage(g_Hwnd, WM_STARTEDITING, 0, (LPARAM) &rightClickedMessage);
+			SendMessage(GetParent(m_hwnd), WM_STARTEDITING, 0, (LPARAM) &rightClickedMessage);
 			break;
 		}
 		case ID_DUMMYPOPUP_COPYTEXT:
@@ -3288,10 +3288,9 @@ void MessageList::HandleRightClickMenuCommand(int command)
 			static char buffer[8192];
 			snprintf(buffer, sizeof buffer, TmGetString(IDS_CONFIRM_DELETE).c_str(), pMsg->m_msg->m_author.c_str(), pMsg->m_msg->m_dateFull.c_str(), pMsg->m_msg->m_message.c_str());
 			LPCTSTR xstr = ConvertCppStringToTString(buffer);
-			if (MessageBox(g_Hwnd, xstr, TmGetTString(IDS_CONFIRM_DELETE_TITLE), MB_YESNO | MB_ICONQUESTION) == IDYES)
-			{
+
+			if (MessageBox(GetParent(m_hwnd), xstr, TmGetTString(IDS_CONFIRM_DELETE_TITLE), MB_YESNO | MB_ICONQUESTION) == IDYES)
 				GetDiscordInstance()->RequestDeleteMessage(m_channelID, rightClickedMessage);
-			}
 
 			free((void*)xstr);
 			break;
@@ -3301,7 +3300,7 @@ void MessageList::HandleRightClickMenuCommand(int command)
 			Snowflake sf[2];
 			sf[0] = pMsg->m_msg->m_snowflake;
 			sf[1] = pMsg->m_msg->m_author_snowflake;
-			SendMessage(g_Hwnd, WM_STARTREPLY, 0, (LPARAM)sf);
+			SendMessage(GetParent(m_hwnd), WM_STARTREPLY, 0, (LPARAM)sf);
 			break;
 		}
 		case ID_DUMMYPOPUP_PINMESSAGE:
@@ -3315,7 +3314,7 @@ void MessageList::HandleRightClickMenuCommand(int command)
 
 			LPCTSTR xstr = ConvertCppStringToTString(buffer);
 
-			if (MessageBox(g_Hwnd, xstr, TmGetTString(IDS_CONFIRM_PIN_TITLE), MB_YESNO | MB_ICONQUESTION) == IDYES)
+			if (MessageBox(GetParent(m_hwnd), xstr, TmGetTString(IDS_CONFIRM_PIN_TITLE), MB_YESNO | MB_ICONQUESTION) == IDYES)
 			{
 				// TODO
 			}
@@ -3391,7 +3390,7 @@ LRESULT CALLBACK MessageList::WndProc(HWND hWnd, UINT uMsg, WPARAM wParam, LPARA
 
 			if (amount != 1)
 			{
-				MessageBox(g_Hwnd, TmGetTString(IDS_CANT_UPLOAD_SEVERAL), TmGetTString(IDS_PROGRAM_NAME), MB_ICONWARNING | MB_OK);
+				MessageBox(GetParent(pThis->m_hwnd), TmGetTString(IDS_CANT_UPLOAD_SEVERAL), TmGetTString(IDS_PROGRAM_NAME), MB_ICONWARNING | MB_OK);
 				DragFinish(hDrop);
 				return 0;
 			}
@@ -3608,7 +3607,7 @@ LRESULT CALLBACK MessageList::WndProc(HWND hWnd, UINT uMsg, WPARAM wParam, LPARA
 			RECT rectParent = {};
 			GetClientRect(pThis->m_hwnd, &rectParent);
 			RECT rectParent2 = {};
-			GetClientRect(g_Hwnd, &rectParent2);
+			GetClientRect(GetParent(pThis->m_hwnd), &rectParent2);
 
 			POINT pt;
 			pt.x = xPos;
