+++ src/windows/TextToSpeech.cpp
@@ -24,7 +24,7 @@ void TextToSpeech::Initialize()
 	{
 		std::string xstr = TmGetString(IDS_FAILED_TO_INITIALIZE_SAPI) + "\n\n(" + std::to_string((long)hr) + ") " + GetStringFromHResult(hr);
 		LPCTSTR tstr1 = ConvertCppStringToTString(xstr);
-		MessageBox(g_Hwnd, tstr1, TmGetTString(IDS_FAILED_TO_INITIALIZE_SAPI_TITLE), MB_OK | MB_ICONERROR);
+		MessageBox(GetMainHWND(), tstr1, TmGetTString(IDS_FAILED_TO_INITIALIZE_SAPI_TITLE), MB_OK | MB_ICONERROR);
 		free((void*)tstr1);
 		// TODO: Let the user know that there is no text to speech.
 		return;
@@ -52,7 +52,7 @@ void TextToSpeech::Speak(const std::string& str)
 	{
 		std::string xstr = TmGetString(IDS_FAILED_TO_SPEAK_MESSAGE) + " \"" + str + "\":\n\n(" + std::to_string((long)hr) + ") " + GetStringFromHResult(hr);
 		LPCTSTR tstr1 = ConvertCppStringToTString(xstr);
-		MessageBox(g_Hwnd, tstr1, TmGetTString(IDS_PROGRAM_NAME), MB_OK | MB_ICONERROR);
+		MessageBox(GetMainHWND(), tstr1, TmGetTString(IDS_PROGRAM_NAME), MB_OK | MB_ICONERROR);
 		free((void*)tstr1);
 	}
 	free((void*)tstr);
